% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_match.R
\name{parse_match}
\alias{parse_match}
\title{Parse Match Function}
\usage{
parse_match(file_path)
}
\arguments{
\item{file_path}{The file path of the match JSON file.}
}
\value{
A list containing the extracted data elements in data frames.
}
\description{
This function parses a match JSON file and extracts various data elements such as locations, grenades, kills, bomb events, round details, damage, and weapon fires.
}
\examples{
file_path <- "C:/Users/shaw_/Documents/Consultancy/Guild/JSON/angels-vs-guild.json"
test <- parse_match(file_path = file_path)

}
