% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position_transform.R
\name{position_transform}
\alias{position_transform}
\title{Position Transform Function}
\usage{
position_transform(map_name, position, axis)
}
\arguments{
\item{map_name}{The name of the map for which the transformation is needed.}

\item{position}{The X or Y coordinate to be transformed.}

\item{axis}{The axis on which the transformation is to be applied, either "x" or "y".}
}
\value{
The transformed X or Y coordinate based on the map name, starting position, and scale factor.
}
\description{
This function transforms X or Y coordinates based on the map name, starting position, and scale factor obtained from a predefined map data.
}
\examples{
position_transform("de_nuke", position = 2300, axis = "x")

}
